Добавил автарку для пользовтелей, разбил app.blade.php на inlcude, сдела боле акрутные карточки товра

+++ database/migrations/2025_09_27_155224_add_icon_to_categories_table.php
@@ -9,14 +9,18 @@ return new class extends Migration
 public function up(): void
 {
     Schema::table('categories', function (Blueprint $table) {
-        $table->string('icon')->nullable()->after('slug'); // путь к иконке
+        if (!Schema::hasColumn('categories', 'icon')) {
+            $table->string('icon')->nullable()->after('slug'); // путь к иконке
+        }
     });
 }
 
 public function down(): void
 {
     Schema::table('categories', function (Blueprint $table) {
-        $table->dropColumn('icon');
+        if (Schema::hasColumn('categories', 'icon')) {
+            $table->dropColumn('icon');
+        }
     });
 }
